+++ CHANGELOG.md
@@ -5,11 +5,17 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-rc.3]
+
+### Removed
+
+- The `valid` option is removed from `message`, any status >= 400 will return a fail.
+
 ## [1.0.0-rc.2]
 
 ### Changed
 
-- The `Validation` type is now called `SuperValidate`.
+- The `Validation` type is now called `SuperValidated`.
 - `StringPath` and `StringPathLeaves` are renamed to `FormPath` and `FormPathLeaves`.
 
 ### Removed